cleanup: dropped old backlight module.
Moved Backlight2 to just Backlight, cleaned up polkit policy conf.
