+++ TODO.md
@@ -1,30 +1,25 @@
-## 5.x
-- [ ] Keep it up to date with possible ddcutil api changes
+## 6.0
 
-### Pipewire
-- [ ] Fix set_camera_setting() impl -> how to get current value? how to set a new value?
-
-### Screen
-- [x] fix `rgb_frame_brightness` impl
-- [x] fix xorg impl
-
-## 6.x (api break release)
+### Backlight
+- [x] Drop old backlight API
+- [x] Move Backlight2 to Backlight
+- [x] Drop {Lower,Raise,Set}All from clightd polkit policy
+- [ ] Add support for X and Wl to Backlight (ie: brightness emulation, like xrandr)
+- - [ ] Xorg: https://stackoverflow.com/questions/5811279/possible-to-change-screen-brightness-with-c
+- - [ ] Wl: https://gitlab.com/chinstrap/gammastep/-/blob/master/src/redshift.c
+- - [ ] Native.{Internal,DDC}
 
 ### Generic
 - [ ] Drop is_smooth options for gamma
 - [ ] avoid returning boolean where it does not make sense
-- [ ] Drop old BACKLIGHT module -> in case, drop {Lower,Raise,Set}All from clightd polkit policy
-- [ ] Rename Backlight2 to Backlight
 
-### Move to user service (?)
-- [ ] move clightd to user service
-- [ ] Drop polkit and use sd_session_is_active() 
-- [ ] Add udev rules for yoctolight, als, keyboard and backlight modules for "clightd" 
-- [ ] Drop useless API params (eg: DISPLAY, XAUTHORITY, XDG_RUNTIME_DIR etc etc)
-- [ ] # groupadd clightd
-- [ ] # usermod -aG clightd myusername
-- [ ] # echo 'KERNEL=="i2c-[0-9]*", GROUP="clightd"' >> /etc/udev/rules.d/10-local_i2c_group.rules
-... would break https://github.com/FedeDP/Clight/issues/144 ...
+## 6.x
+
+### Backlight
+- [ ] Keep it up to date with possible ddcutil api changes
+
+### Pipewire
+- [ ] Fix set_camera_setting() impl -> how to get current value? how to set a new value?
 
 ## Ideas
 - [ ] follow ddcci kernel driver and in case, drop ddcutil and add the kernel driver as clightd opt-dep
